# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apple Inc. API Management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/list/{page}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated list of products",
                "operationId": "0b437b3b059896eb544343227f15fe5d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "The page number to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Retrieved Successfully."
                                        },
                                        "totalrecords": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totpage": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Products is empty"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/products/search/{key}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products by description",
                "operationId": "9b9ba70475c81ec70f82797285b6c73a",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Search keyword for descriptions",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Searched found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Searched found.."
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "descriptions": {
                        "type": "string",
                        "example": "Gaming Laptop"
                    },
                    "qty": {
                        "type": "integer",
                        "example": 10
                    },
                    "unit": {
                        "type": "string",
                        "example": "pcs"
                    },
                    "costprice": {
                        "type": "number",
                        "format": "float",
                        "example": 800
                    },
                    "sellprice": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "saleprice": {
                        "type": "number",
                        "format": "float",
                        "example": 950
                    },
                    "productpicture": {
                        "type": "string",
                        "example": "laptop.jpg"
                    },
                    "alertstocks": {
                        "type": "integer",
                        "example": 5
                    },
                    "criticalstocks": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}